E3G/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
distro==1.9.0
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
openai==1.35.1
openpyxl==3.1.4
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.3.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pypdfium2==4.30.0
python-docx==1.1.2
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
